JavaFX Stats and Apache POI Stats
Started wokring on interacting with Microsoft Excel spreadsheets with my program. Will most likely be used to write information and not read it.

Changed Exercise class to include max weight, reps, volume, and one rep max estimate. Will use these on the graphs in the future.

Changed reps instance variable in Set class from int to double. Needed to do this in order to perform the division required for one rep max estimation

+++ FitnessAppJavaFX/src/model/Set.java
@@ -10,7 +10,7 @@ package model;
  */
 public class Set {
 
-	int reps;
+	double reps;
 	double weight;
 
 	/**
@@ -37,7 +37,7 @@ public class Set {
 	 * 
 	 * @return reps
 	 */
-	public int getReps() {
+	public double getReps() {
 		return this.reps;
 	}
 
@@ -73,7 +73,7 @@ public class Set {
 	 * 
 	 * @return
 	 */
-	public int increaseReps() {
+	public double increaseReps() {
 		this.reps += 1;
 		return this.reps;
 	}
@@ -83,7 +83,7 @@ public class Set {
 	 * 
 	 * @return reps
 	 */
-	public int decreaseReps() {
+	public double decreaseReps() {
 		this.reps -= 1;
 		return this.reps;
 	}
